refactor(harness,cli)!: remove custom provenance persistence
Change E of the harness-integration graph: the cli bus adapter has been
the live ArtifactRegistry realization since the provenance bridge landed
(D/D2/D3), so the harness's own persistence path is dead code whose
header misleads readers toward a format nothing reads back.

Harness:
- delete FilesystemArtifactRegistry and the provenance-index.json format;
  createNoopArtifactRegistry becomes the seam's OSS realization (BREAKING:
  the barrel drops createFilesystemArtifactRegistry/FilesystemArtifactRegistryDeps)
- delete the dead workspace write-snapshot seam (recordSnapshot) and the
  plumbing it justified (mutator logger machinery, unused runId dep);
  the step-scoped lineage collector stays the single home for tool-write
  lineage if it is ever wired
- delete the consumer-less isToolNoisePath export
- fix stale comments (processProvenanceFrame refs, the phantom
  StepMetadata.sourceRunIds box in classifyReadPath) and document
  ProvenanceFrame.deletes as reserved on the wire
- sync spec deltas: harness-durable-runtime (noop seam realization),
  artifact-manifest (noop citations), harness-workspace-tools (snapshot
  requirement removed), explicit-input-classification (prior-run comment
  mandate made truthful)

CLI (spec files only, no source changes): prov-signing/prov-chain no
longer describe a degrade-to-unsigned flush the code never had; the
migration narrative matches the single-baseline v1 schema incl.
provenance_prev_chain_hash; prov-verify lists all 8 VerifyResult
variants; export is documented as hard-failing when signing is
impossible.

Gates: harness tsc clean; bun test 726 tests, 0 fail; cli specs
openspec-validate clean.

OpenSpec change: remove-custom-provenance-persistence (archived).
