+++ harness/CONTEXT.md
@@ -47,7 +47,7 @@ Each seam is an interface the harness declares; an embedder binds one realizatio
 
 - **`RunAuthorizer`** (`execution/run-authorizer.ts`) — turns the opaque caller `auth` into a durable `RunSession` at the async edge. The sole constructor of `RunSession`, so it is the single chokepoint where in-process identity becomes durable workflow identity. OSS realization: `auth/local-run-authorizer.ts` issues a `RunSession` with no remote mint and no revoke — authorization is purely structural.
 - **`ResolveBilling`** (`billing/resolver.ts`) — returns a ready-to-attach call-attribution header map the provider spreads at the LLM/embedding call site; resolved lazily, only at the wire boundary. OSS realization: `createNoopBillingResolver` returns `{}` (empty headers).
-- **`ArtifactRegistry`** (`execution/artifact-registry.ts`) — post-step artifact recording (content-attested lineage): `register` the step's manifest + `sync` its bytes to permanent storage, both session-scoped (the adapter authenticates per-run off the session). OSS realization: `FilesystemArtifactRegistry` — a real local provenance index; `sync` is a no-op (bytes already live in the local session tree).
+- **`ArtifactRegistry`** (`execution/artifact-registry.ts`) — post-step artifact recording (content-attested lineage): `register` the step's manifest + `sync` its bytes to permanent storage, both session-scoped (the adapter authenticates per-run off the session). OSS realization: `createNoopArtifactRegistry` — registers nothing externally and reports zero failures (the local `cortex_artifacts` ledger is written by the harness around the seam); `sync` is a no-op (bytes already live in the local session tree).
 - **`RunCharge`** (`billing/run-charge.ts`) — the run-level billing bracket `executeAnalysis` opens at init and closes on the terminal path (one of four reasons). Local realization: `createNoopRunCharge` — no-op.
 - **`PreviewPublisher`** (`tools/report/preview-publisher.ts`) — report preview URL publishing. Local realization: `UnavailablePreviewPublisher` — reports still build; preview URL minting is gated.
 - **`RunLauncher`** (`execution/run-launcher.ts`) — starts a registered workflow under a caller-chosen id (`launch` fire-and-forget, `launchAndAwait` inline with cancel-on-abort, cancellation hidden behind a discriminated `LaunchOutcome`). The DBOS quarantine seam: it is why `execute_plan` / `run_ephemeral` do not import the durability engine. Single host-neutral realization `createDbosRunLauncher` (`execution/dbos-run-launcher.ts`) shared by every embedder.
