+++ harness/openspec/changes/archive/2026-07-07-remove-custom-provenance-persistence/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: spec-driven
+created: 2026-07-07
