+++ harness/openspec/specs/artifact-manifest/spec.md
@@ -167,7 +167,7 @@ return all-zero counts immediately and SHALL NOT call the registry.
 as fail-fast: a non-zero `externalFailed` SHALL throw with the
 per-file failure detail, and the sandbox-step body SHALL tear down the sandbox,
 mark the step failed, and re-raise so the parent's fail-fast cascade fires. The
-OSS `FilesystemArtifactRegistry` returns `externalFailed: 0` and never trips
+OSS `createNoopArtifactRegistry` returns `externalFailed: 0` and never trips
 this. The enrichment stages — file-metadata generation, step-summary generation,
 and vector indexing — SHALL run under `safeRun`/`safeRunValue` so any single
 failure degrades without failing the step.
@@ -200,9 +200,9 @@ existing row rather than fail on the duplicate key.
 
 `cortex_artifacts` SHALL track external sync state via `artifact_id` (set after
 external registration by `updateArtifactId`) and `file_id` (set when an adapter
-reports an external file identity, applied in batch by `updateFileIds`). The
-filesystem registry MAY leave both null because the bytes already live in the
-local session tree. `queryUnsyncedStepArtifacts(pool, resourceId, runId,
+reports an external file identity, applied in batch by `updateFileIds`). A
+registry realization with no external system (the OSS noop) MAY leave both null
+because the bytes already live in the local session tree. `queryUnsyncedStepArtifacts(pool, resourceId, runId,
 stepId)` SHALL return step-output rows where `artifact_id IS NOT NULL AND file_id
 IS NULL`, ordered by `created_at`.
