+++ harness/openspec/specs/harness-workspace-tools/spec.md
@@ -419,39 +419,3 @@ top of a resolved, scope-checked path.
 - **THEN** both surfaces SHALL resolve to the same session-tree location and the
   content written SHALL be returned
 
-### Requirement: Mutate surface records SHA-256 content snapshots for provenance
-
-Every successful `write_file` and `edit_file` invocation SHALL feed a SHA-256
-content snapshot of the written bytes to the provenance collector keyed by `{
-analysisId, runId, stepId, path, sha256, bytes, timestamp }`, captured at the
-mutation site (not reconstructed later by re-hashing artifact files). Artifact
-registration consumes these snapshots as lineage without re-computing the hash.
-
-`execute_command`'s in-sandbox writes SHALL NOT be hashed by the harness — they
-are captured by sandbox-server's provenance frames (see the
-sandbox-provenance-tracking spec) and reconciled by artifact registration. A
-provenance recording failure SHALL NOT fail the mutate tool: the write is the
-durable outcome, the snapshot is best-effort metadata logged via the standard
-logger, and the tool SHALL still return its normal `ok` data variant.
-
-#### Scenario: A write_file call records a snapshot with the file's SHA-256
-
-- **GIVEN** a `write_file` call writing N bytes of known content C
-- **WHEN** the write completes successfully
-- **THEN** the provenance collector SHALL receive a record whose `path` matches
-  the resolved workspace path, `bytes` equals N, and `sha256` equals
-  `sha256Hex(C)`
-
-#### Scenario: An edit_file call records a snapshot of the post-edit content
-
-- **GIVEN** an `edit_file` call transforming content from C1 to C2
-- **WHEN** the edit completes successfully
-- **THEN** the provenance collector SHALL receive a snapshot whose `sha256`
-  equals `sha256Hex(C2)`, not `sha256Hex(C1)`
-
-#### Scenario: A failing provenance recording does not fail the write
-
-- **GIVEN** a provenance collector that throws when called
-- **WHEN** `write_file` records a snapshot after a successful write
-- **THEN** the tool result SHALL still be an `ok` data variant and `execute`
-  SHALL NOT throw
